Ensure context update before answering requests.
This fixes two major (and related) bugs when processing client requests
just after the source code has been changed, and drastically improves
the behavior of the server when the source code is edited quickly.

1. Just after the source code is edited, the client sends requests such
as "document symbols" or "code lens", but the server did not wait for
the context to be updated before processing these requests. So the
request was processed with an outdated context, and the response
was generally incorrect.
The fix is to wait for the context to be updated before processing the
request, which means that these requests are now processed
asynchronously (and do not slow down the context update).

2. Related to this, when the source code is edited again before the
server has answered a pending request (which happens now that
requests are processed asynchronously), the client will cancel the
previous request with a notification, and send a new one.
This is now handled properly: the pending request identifiers are
tracked to avoid processing cancelled requests.