Add support for multiple build targets.
Each Meson target is now created explicitly with the corresponding
source files, with the restriction that only one executable target may
be active at any one time (to avoid multiple symbol definitions).

When the server receives a request or notification for an inactive
source file, the corresponding build target is activated and the
previously active executable target is deactivated, then the compiler
context is updated.

This should allow to handle the case of one common library and
several executables (for tests etc.) transparently (in most cases).