Fix nasty reentrancy issue.
When using the synchronous JsonRpc.Client methods, the thread can be
preempted by the library to process inbound requests.

Here the issue is that completion requests could be processed while the
diagnostics are still being sent causing a crash because the context
is rebuilt by the completion request (failed modified collection under
iterator assertion).

Using the async version seems to fix this issue (another possibility would be
to prepare the list of diagnostics to send in advance and send them in one
go at the end.